+++ shared/reducers/plan-billing.js
@@ -1,8 +1,9 @@
 // @flow
 import * as CommonConstants from '../constants/common'
 import * as Constants from '../constants/plan-billing'
+import * as Types from '../constants/types/plan-billing'
 
-const initialState: Constants.State = {
+const initialState: Types.State = {
   availablePlans: null,
   errorMessage: null,
   paymentInfo: null,
@@ -11,50 +12,51 @@ const initialState: Constants.State = {
 }
 
 export default function(
-  state: Constants.State = initialState,
-  action: Constants.Actions | {type: 'common:resetStore', payload: void}
-): Constants.State {
+  state: Types.State = initialState,
+  // TODO gen and type this if we actually use this thing
+  action: any
+): Types.State {
   switch (action.type) {
     case CommonConstants.resetStore:
       return {...initialState}
-    // case Constants.updateBillingAndQuota:
-    // if (action.error) {
-    // console.warn('Error in action: ', action)
-    // return state
-    // }
-    // return {
-    // ...state,
-    // ...action.payload,
-    // }
-    // case Constants.updateAvailablePlans:
-    // if (action.error) {
-    // console.warn('Error in action: ', action)
-    // return state
-    // }
-    // return {
-    // ...state,
-    // ...action.payload,
-    // }
-    // case Constants.updatePaymentInfo:
-    // if (action.error) {
-    // console.warn('Error in action: ', action)
-    // return state
-    // }
-    // return {
-    // ...state,
-    // ...action.payload,
-    // }
-    // case Constants.billingError:
-    // if (action.error) {
-    // return {
-    // ...state,
-    // errorMessage: action.payload.errorText,
-    // }
-    // }
-    // return {
-    // ...state,
-    // errorMessage: null,
-    // }
+    case Constants.updateBillingAndQuota:
+      if (action.error) {
+        console.warn('Error in action: ', action)
+        return state
+      }
+      return {
+        ...state,
+        ...action.payload,
+      }
+    case Constants.updateAvailablePlans:
+      if (action.error) {
+        console.warn('Error in action: ', action)
+        return state
+      }
+      return {
+        ...state,
+        ...action.payload,
+      }
+    case Constants.updatePaymentInfo:
+      if (action.error) {
+        console.warn('Error in action: ', action)
+        return state
+      }
+      return {
+        ...state,
+        ...action.payload,
+      }
+    case Constants.billingError:
+      if (action.error) {
+        return {
+          ...state,
+          errorMessage: action.payload.errorText,
+        }
+      }
+      return {
+        ...state,
+        errorMessage: null,
+      }
   }
   return state
 }
